Currency controller/service/repository changed

+++ src/main/java/by/weekmenu/api/repository/CurrencyRepository.java
@@ -3,5 +3,5 @@ package by.weekmenu.api.repository;
 import by.weekmenu.api.entity.Currency;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CurrencyRepository extends CrudRepository<Currency, Byte>  {
+public interface CurrencyRepository extends CrudRepository<Currency, Integer>  {
 }
